+++ packages/components/src/synergy.ts
@@ -4,6 +4,7 @@
 // !!! and run `pnpm build` to recreate this file
 
 // Components
+export { default as SynAccordion } from './components/accordion/accordion.js';
 export { default as SynAlert } from './components/alert/alert.js';
 export { default as SynBadge } from './components/badge/badge.js';
 export { default as SynBreadcrumbItem } from './components/breadcrumb-item/breadcrumb-item.js';
@@ -12,6 +13,7 @@ export { default as SynButtonGroup } from './components/button-group/button-grou
 export { default as SynButton } from './components/button/button.js';
 export { default as SynCard } from './components/card/card.js';
 export { default as SynCheckbox } from './components/checkbox/checkbox.js';
+export { default as SynDetails } from './components/details/details.js';
 export { default as SynDialog } from './components/dialog/dialog.js';
 export { default as SynDivider } from './components/divider/divider.js';
 export { default as SynDrawer } from './components/drawer/drawer.js';